+++ sdk/nodejs/helm/v2/helm.ts
@@ -218,7 +218,7 @@ export class Chart extends yaml.CollectionComponentResource {
                     },
                 ).toString();
                 return this.parseTemplate(
-                    yamlStream, cfg.transformations, cfg.resourcePrefix, configDeps, cfg.namespace, opts);
+                    yamlStream, cfg.transformations, cfg.resourcePrefix, configDeps, cfg.namespace);
             } catch (e) {
                 // Shed stack trace, only emit the error.
                 throw new pulumi.RunError(e.toString());
@@ -236,20 +236,12 @@ export class Chart extends yaml.CollectionComponentResource {
         resourcePrefix: string | undefined,
         dependsOn: pulumi.Resource[],
         defaultNamespace: string | undefined,
-        opts?: pulumi.ComponentResourceOptions,
     ): pulumi.Output<{ [key: string]: pulumi.CustomResource }> {
-        // Rather than using the default provider for the following invoke call, determine the
-        // provider from the parent if specified, or fallback to using the version specified
+        // Rather than using the default provider for the following invoke call, use the version specified
         // in package.json.
-        let invokeOpts: pulumi.InvokeOptions = {async: true};
-        if (opts?.parent) {
-            invokeOpts = {...invokeOpts, parent: opts.parent};
-        } else {
-            invokeOpts = {...invokeOpts, version: getVersion()};
-        }
+        let invokeOpts: pulumi.InvokeOptions = { async: true, version: getVersion() };
 
-        const promise = pulumi.runtime.invoke(
-            "kubernetes:yaml:decode", {text, defaultNamespace}, invokeOpts);
+        const promise = pulumi.runtime.invoke("kubernetes:yaml:decode", {text, defaultNamespace}, invokeOpts);
         return pulumi.output(promise).apply<{[key: string]: pulumi.CustomResource}>(p => yaml.parse(
             {
                 resourcePrefix: resourcePrefix,