+++ tests/examples/python/python_test.go
@@ -85,7 +85,7 @@ func TestYaml(t *testing.T) {
 		ExpectRefreshChanges: true,
 		ExtraRuntimeValidation: func(t *testing.T, stackInfo integration.RuntimeValidationStackInfo) {
 			assert.NotNil(t, stackInfo.Deployment)
-			assert.Equal(t, 15, len(stackInfo.Deployment.Resources))
+			assert.Equal(t, 14, len(stackInfo.Deployment.Resources))
 
 			sort.Slice(stackInfo.Deployment.Resources, func(i, j int) bool {
 				ri := stackInfo.Deployment.Resources[i]
@@ -159,11 +159,9 @@ func TestYaml(t *testing.T) {
 			// Verify the provider resources.
 			provRes := stackInfo.Deployment.Resources[12]
 			assert.True(t, providers.IsProviderType(provRes.URN.Type()))
-			provRes = stackInfo.Deployment.Resources[13]
-			assert.True(t, providers.IsProviderType(provRes.URN.Type()))
 
 			// Verify root resource.
-			stackRes := stackInfo.Deployment.Resources[14]
+			stackRes := stackInfo.Deployment.Resources[13]
 			assert.Equal(t, resource.RootStackType, stackRes.URN.Type())
 
 			// TODO[pulumi/pulumi#2782] Testing of secrets blocked on a bug in Python support for secrets.
@@ -312,7 +310,7 @@ func TestHelm(t *testing.T) {
 		ExpectRefreshChanges: true, // PodDisruptionBudget status gets updated by the Deployment.
 		ExtraRuntimeValidation: func(t *testing.T, stackInfo integration.RuntimeValidationStackInfo) {
 			assert.NotNil(t, stackInfo.Deployment)
-			assert.Equal(t, 16, len(stackInfo.Deployment.Resources))
+			assert.Equal(t, 15, len(stackInfo.Deployment.Resources))
 
 			sort.Slice(stackInfo.Deployment.Resources, func(i, j int) bool {
 				ri := stackInfo.Deployment.Resources[i]